refactor: extract host process primitives
