+++ src/platforms/apple/core/runner/runner-adoption.ts
@@ -1,7 +1,7 @@
 import path from 'node:path';
 import { resolveIosSimulatorDeviceSetPath } from '../../../../utils/device-isolation.ts';
 import { emitDiagnostic } from '../../../../utils/diagnostics.ts';
-import { isProcessAlive } from '../../../../utils/process-identity.ts';
+import { isProcessAlive } from '../../../../utils/host-process.ts';
 import { parseBooleanLiteral } from '../../../../utils/source-value.ts';
 import type { DeviceInfo } from '../../../../kernel/device.ts';
 import type { ExecResult } from '../../../../utils/exec.ts';
